Added Redis cache to the application to improve the performance

+++ src/main/java/com/insulin/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.insulin.controller;
 
+import com.insulin.exception.model.InvalidDataException;
 import com.insulin.exception.model.UserNotFoundException;
 import com.insulin.model.User;
 import com.insulin.service.EmailService;
@@ -11,6 +12,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import javax.mail.MessagingException;
@@ -28,16 +31,30 @@ public class UserController {
         this.emailService = emailService;
     }
 
-    @DeleteMapping("/user/{id}")
-    public ResponseEntity<HttpResponse> deleteUserById(@PathVariable("id") Long id) throws UserNotFoundException, MessagingException {
+    @DeleteMapping("/delete/{id}")
+    @PreAuthorize("hasAnyAuthority('PATIENT', 'MEDIC', 'ADMIN')")
+    public ResponseEntity<HttpResponse> deleteUserById(@PathVariable("id") Long id, Authentication auth)
+            throws UserNotFoundException, MessagingException, InvalidDataException {
+        String principal = (String) auth.getPrincipal();
         User currentUser = this.userService.getUserById(id);
         if (currentUser == null) {
             logger.error("The provided id is invalid. User not found");
             throw new UserNotFoundException("The provided id does not map to any user");
         }
+        if (!principal.equals(currentUser.getUsername())) {
+            logger.error("A user could not delete another user account!");
+            throw new InvalidDataException("The id does not match to the current account!");
+        }
+
         this.emailService.sendDeleteEmail(currentUser.getDetails().getFirstName(), currentUser.getDetails().getEmail());
         logger.info("Delete message has been emitted");
         this.userService.deleteUser(id);
         return HttpResponseUtils.buildHttpResponseEntity(HttpStatus.OK, "User deleted with success");
+    }
+
+    @GetMapping("/{id}")
+    @PreAuthorize("hasAnyAuthority('PATIENT', 'MEDIC', 'ADMIN')")
+    public ResponseEntity<User> getUserById(@PathVariable("id") Long id) throws UserNotFoundException {
+        return new ResponseEntity<>(userService.getUserById(id), HttpStatus.OK);
     }
 }
